Delete lesson's folder when delete lesson

+++ backend/src/main/java/com/mjsasha/lesdoc/services/LessonsService.java
@@ -36,10 +36,10 @@ public class LessonsService {
     }
 
     public void delete(Integer id) {
-        try {
-            lessonsRepository.deleteById(id);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        Lesson lesson = read(id);
+        if (lesson == null) throw new EntityNotFoundException("Entity with id-" + id + " not exist.");
+
+        fileStorageService.removeDirectory(lesson.getFolderName());
+        lessonsRepository.deleteById(id);
     }
 }
